+++ src/base-command.ts
@@ -60,18 +60,12 @@ export default abstract class BaseCommand extends Command {
     try {
       this.sentry = await SentryService.create(this.app, this.constructor as any);
     } catch (e) {
-      // nothing to do
+      // dont fail when adding metadata
     }
   }
 
   async finally(_: Error | undefined): Promise<any> {
-    if (this.app.config.environment !== ENVIRONMENT.TEST && !_) {
-      const { filtered_sentry_args, filtered_sentry_flags } = await this._getNonSensitiveSentryMetadata();
-      await this.sentry.setScopeExtra('command_args', filtered_sentry_args);
-      await this.sentry.setScopeExtra('command_flags', filtered_sentry_flags);
-      return await this.sentry.endSentryTransaction();
-    }
-    return super.finally(_);
+    return await this.endSentryTransaction();
   }
 
   // Move all args to the front of the argv to get around: https://github.com/oclif/oclif/issues/190
@@ -114,20 +108,16 @@ export default abstract class BaseCommand extends Command {
   async catch(err: any): Promise<void> {
     if (err.oclif && err.oclif.exit === 0) return;
 
-    if (err.stack) {
-      err.stack = [...new Set(err.stack.split('\n'))].join("\n");
-    }
-
-    const { filtered_sentry_args, filtered_sentry_flags } = await this._getNonSensitiveSentryMetadata();
-    await this.sentry.setScopeExtra('command_args', filtered_sentry_args);
-    await this.sentry.setScopeExtra('command_flags', filtered_sentry_flags);
-    await this.sentry.endSentryTransaction(err);
-
-    if (err instanceof ValidationErrors) {
-      return prettyValidationErrors(err);
-    }
-
     try {
+
+      if (err.stack) {
+        err.stack = [...new Set(err.stack.split('\n'))].join("\n");
+      }
+
+      if (err instanceof ValidationErrors) {
+        return prettyValidationErrors(err);
+      }
+
       if (err.response?.data instanceof Object) {
         err.message += `\nmethod: ${err.config.method}`;
         for (const [k, v] of Object.entries(err.response.data)) {
@@ -150,8 +140,11 @@ export default abstract class BaseCommand extends Command {
     } catch {
       this.warn('Unable to add more context to error message');
     }
-
-    return super.catch(err);
+    const app_env = this.app?.config?.environment;
+    if (!this.sentry || !app_env || app_env === ENVIRONMENT.TEST) {
+      return super.catch(err);
+    }
+    return await this.endSentryTransaction(err);
   }
 
   async _filterNonSensitiveSentryMetadata(non_sensitive: Set<string>, metadata: any): Promise<any> {
@@ -160,7 +153,12 @@ export default abstract class BaseCommand extends Command {
       .map(key => ({ [key[0]]: key[1] }));
   }
 
-  async _getNonSensitiveSentryMetadata(_args?: any[], _flags?: any[]): Promise<any> {
+  async endSentryTransaction(err?: any): Promise<any> {
+    const app_env = this.app?.config?.environment;
+    if (!this.sentry || !app_env || app_env === ENVIRONMENT.TEST) {
+      return err;
+    }
+
     const calling_class = this.constructor as any;
 
     const non_sensitive = new Set([
@@ -170,11 +168,21 @@ export default abstract class BaseCommand extends Command {
 
     try {
       const { args, flags } = await this.parse(calling_class);
-      const filtered_sentry_args = await this._filterNonSensitiveSentryMetadata(non_sensitive, _args || args);
-      const filtered_sentry_flags = await this._filterNonSensitiveSentryMetadata(non_sensitive, _flags || flags);
-      return { filtered_sentry_args, filtered_sentry_flags };
+      const filtered_sentry_args = await this._filterNonSensitiveSentryMetadata(non_sensitive, args);
+      const filtered_sentry_flags = await this._filterNonSensitiveSentryMetadata(non_sensitive, flags);
+
+      await this.sentry.setScopeExtra('command_args', filtered_sentry_args);
+      await this.sentry.setScopeExtra('command_flags', filtered_sentry_flags);
     } catch {
-      return { filtered_sentry_args: [], filtered_sentry_flags: []};
+      this.warn('Failed to get command metadata');
     }
+
+    await this.sentry.endSentryTransaction(err);
+
+    if (!err) {
+      return await super.finally(err);
+    }
+    return await super.catch(err);
   }
+
 }