+++ src/common/utils/docker-buildx.utils.ts
@@ -48,12 +48,12 @@ export default class DockerBuildXUtils {
 
   public static async getBuilder(config: config): Promise<string> {
     const is_local = this.isLocal(config);
-    const builder = is_local ? 'architect-local' : 'architect';
+    const builder = (is_local ? 'architect-local' : 'architect') + `-${process.platform}`;
 
     // Create a docker context
     try {
       await docker(['context', 'create', `${builder}-context`]);
-    // eslint-disable-next-line no-empty
+      // eslint-disable-next-line no-empty
     } catch (err) { }
 
     try {
@@ -71,7 +71,7 @@ export default class DockerBuildXUtils {
           stdio: 'inherit',
         });
       }
-    // eslint-disable-next-line no-empty
+      // eslint-disable-next-line no-empty
     } catch { }
 
     return builder;
@@ -81,7 +81,7 @@ export default class DockerBuildXUtils {
     if (use_console) {
       process.stdin.setRawMode(true);
     }
-    const cmd = execa('docker', [`--context=${docker_builder_name}-context`, 'buildx', ...args], execa_opts);
+    const cmd = execa('docker', [`--context=${docker_builder_name}-context`, 'buildx', '--builder', docker_builder_name, ...args], execa_opts);
     if (use_console) {
       cmd.on('exit', () => {
         process.exit();