+++ test/utils/mocks.ts
@@ -1,11 +1,13 @@
 import { test } from '@oclif/test';
 import path from 'path';
 import AuthClient from '../../src/app-config/auth';
+import SecretUtils from '../../src/architect/secret/secret.utils';
 import { DockerComposeUtils } from '../../src/common/docker-compose';
-import DockerBuildXUtils from '../../src/common/utils/docker-buildx.utils';
 import * as Docker from '../../src/common/utils/docker';
+import DockerBuildXUtils from '../../src/common/utils/docker-buildx.utils';
 
 export const MOCK_API_HOST = 'http://mock.api.localhost';
+export const MOCK_REGISTRY_HOST = 'http://mock.registry.localhost';
 
 export const TMP_DIR = path.join(__dirname, '../tmp')
 
@@ -29,4 +31,6 @@ export const mockArchitectAuth = test
   .stub(DockerBuildXUtils, 'writeBuildkitdConfigFile', () => { })
   .stub(DockerBuildXUtils, 'dockerBuildX', () => { })
   .stub(DockerBuildXUtils, 'getBuilder', () => { })
+  .stub(SecretUtils, 'getSecrets', () => [])
+  .stub(SecretUtils, 'batchUpdateSecrets', () => [])
   .stub(DockerBuildXUtils, 'convertToBuildxPlatforms', () => { })